Fixes by JOHNY

+++ src/GUI/TF/TFColorVector.h
@@ -273,7 +273,7 @@ public:
 
 		tfAssert(dimensionSizes.size() == 1);
 
-		int inputSize = this->vector_.size();
+		int inputSize = static_cast<int>(this->vector_.size());
 		int outputSize = dimensionSizes[0];
 
 		float oiRatio = outputSize/(float)inputSize;
